requirements.txt
Django
djangorestframework
Pillow
google-generativeai
python-decouple
psycopg2-binary
requests
gunicorn
whitenoise